requirements.txt
autogen-agentchat==0.7.5
autogen-ext[openai]==0.7.5
python-dotenv==1.2.2
bs4==0.0.2